ml/claims_text_pipeline/requirements.txt
fastapi>=0.95,<1.0
pytesseract>=0.3.10
Pillow>=9.5.0
PyMuPDF>=1.22.5
python-docx>=0.8.11
pandas>=2.0.0
textblob>=0.17.1
regex>=2023.6.3
scikit-learn>=1.4
joblib>=1.3
streamlit>=1.35
matplotlib>=3.8
numpy>=1.26
# xgboost and shap are optional; training falls back if unavailable
xgboost>=2.0; python_version>='3.9'
shap>=0.44; python_version>='3.9'
